Function genesis::import_chain

pub async fn import_chain<V: ProofVerifier, DB>(
    sm: &Arc<StateManager<DB>>, 
    path: &str, 
    validate_height: Option<i64>, 
    skip_load: bool
) -> Result<(), Box<dyn Error>> where
    DB: BlockStore + Send + Sync + 'static, 

Import a chain from a CAR file. If the snapshot boolean is set, it will not verify the chain state and instead accept the largest height as genesis.